% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{sc_bulk_p}
\alias{sc_bulk_p}
\alias{sc_bulk_g}
\title{Bulk}
\usage{
sc_bulk_p(urls)

sc_bulk_g(id, force = FALSE)
}
\arguments{
\item{urls}{List or vector of URLS to process.}

\item{id}{id of bulk to retrieve, as returned by \code{\link{sc_bulk_p}}.}

\item{force}{Set this parameter to \code{TRUE} force results to return
even if not all URLs have completed processing.}
}
\description{
Post a large number of URLs all at once to calculate bulk social counts.
}
\examples{
\dontrun{
urls <- c(
  "https://cran.r-project.org/",
  "https://www.rstudio.com/"
)

sc_bulk_p(urls)
}

}
\seealso{
\code{\link{sc_set}} to setup your session.
}
