% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noparam.R
\name{sc_domains}
\alias{sc_domains}
\alias{sc_whitelist}
\title{Domain Whitelist}
\usage{
sc_domains()

sc_whitelist()
}
\description{
Return a list of domains added to your domain whitelist,
and whether the domain whitelist is currently being enforced.
}
\examples{
\dontrun{sc_domains()}

}
\seealso{
\code{\link{sc_set}} to setup your session.
}
