% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{sc_count}
\alias{sc_count}
\title{Count}
\usage{
sc_count(url, ...)
}
\arguments{
\item{url}{URL to return counts.}

\item{...}{Any other options.}
}
\description{
Return counts for a URL.
}
\examples{
\dontrun{sc_count("https://cran.r-project.org/")}

}
\seealso{
\code{\link{sc_set}} to setup your session.
}
