% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set}
\alias{set}
\alias{sc_set}
\alias{sharedcount_key}
\title{Setup session}
\usage{
sc_set(key, quiet = !interactive())

sharedcount_key(key, quiet = !interactive())
}
\arguments{
\item{key}{Your API key.}

\item{quiet}{Whether to print helpful messages to the console.}
}
\description{
Setup your sharedcount session.
}
\examples{
sharedcount_key("xXXXXXxxxXXxxX")

}
