% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noparam.R
\name{sc_quota}
\alias{sc_quota}
\alias{sc_usage}
\alias{sc_status}
\title{Quota}
\usage{
sc_quota()

sc_usage()

sc_status()
}
\description{
Get information on daily quota allocation, historical usage, and API status.
}
\examples{
\dontrun{sc_quota()}

}
\seealso{
\code{\link{sc_set}} to setup your session.
}
